neaten up files
