+++ src/components/Pages/appts.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
@@ -9,13 +8,7 @@ import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
 import Container from '@material-ui/core/Container';
 import Typography from '@material-ui/core/Typography';
-import appointments from './appointments.json'
-
-const useStyles = makeStyles({
-  table: {
-    fontSize: 30
-  },
-});
+import appointments from '../Data/appointments.json'
 
 
 const MyApp = () => {
