+++ src/components/Pages/login.js
@@ -1,20 +1,7 @@
-import React, { useState, handleClick } from "react";
-import Avatar from "@material-ui/core/Avatar";
-import Button from "@material-ui/core/Button";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import TextField from "@material-ui/core/TextField";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Checkbox from "@material-ui/core/Checkbox";
-import Link from "@material-ui/core/Link";
-import Grid from "@material-ui/core/Grid";
-import Box from "@material-ui/core/Box";
-import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
-import Typography from "@material-ui/core/Typography";
+import React, { useState } from "react";
+import { Button, TextField, FormControlLabel, Link, Grid, Box, Typography, Container } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormHelperText from '@material-ui/core/FormHelperText';
-import users from "./users.json";
+import users from "../Data/users.json";
 import Cookies from "universal-cookie";
 
 function Copyright() {
@@ -40,13 +27,9 @@ const useStyles = makeStyles(theme => ({
     flexDirection: "column",
     alignItems: "center"
   },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main
-  },
   form: {
     width: "100%", // Fix IE 11 issue.
-    marginTop: theme.spacing(1)
+    marginTop: theme.spacing(2)
   },
   submit: {
     margin: theme.spacing(3, 0, 2)
@@ -62,7 +45,7 @@ export default function SignIn() {
   function handleSubmit(event) {
     var count=0;
     for (var i = 0; i < users.length; i++){
-      if (users[i].email == email){
+      if (users[i].email === email){
         count++;
         var dbpass=users[i].password
         var verify = bcrypt.compareSync(password, dbpass);
@@ -102,12 +85,13 @@ export default function SignIn() {
             path: "/" ,
             expires: expireDate
           });
+          window.location.reload();
         }
         break;
       
       }
     }
-    if(count==0){
+    if(count===0){
       alert("user not found");
     }
   }
@@ -118,12 +102,11 @@ export default function SignIn() {
 
   return (
     <Container component="main" maxWidth="xs">
-      <CssBaseline />
       <div className={classes.paper}>
         <Typography component="h1" variant="h5">
           Sign in
         </Typography>
-        <form className={classes.form} noValidate>
+        <form className={classes.form}>
           <TextField
             variant="outlined"
             margin="normal"
@@ -134,6 +117,7 @@ export default function SignIn() {
             name="email"
             autoComplete="email"
             autoFocus
+            form className={classes.form}
             onChange={e => setEmail(e.target.value)}
           />
           <TextField
@@ -145,12 +129,14 @@ export default function SignIn() {
             label="Password"
             type="password"
             id="password"
+            form className={classes.form}
             autoComplete="current-password"
             onChange={e => setPassword(e.target.value)}
           />
           <Button
-            type="submit"
+            margin="normal"
             fullWidth
+            form className={classes.form}
             variant="contained"
             color="primary"
             onClick={handleSubmit}
@@ -158,7 +144,7 @@ export default function SignIn() {
           >
             Sign In
           </Button>
-          <Grid container>
+          <Grid container form className={classes.form}>
             <Grid item xs>
               <Link href="/forgot_password" variant="body2">
                 Forgot password?
