+++ src/components/Routes/Routes.js
@@ -1,16 +1,13 @@
 import React from "react";
-import Login from "../login.js";
-import Menu from "../appbar.js";
-import ForgotPassword from "../resetPass.js";
-import SignUp from "../signup.js";
-import SignUpSupporter from "../signUpSupporter.js";
-import tos from "../TOS.js";
-import appbar from "../appbar.js";
-import Home from "../home.js";
-import matchDemo from "../Match.js";
-import appts from "../appts.js";
-import Account from "../account.js";
-import Matching from "../search-appointment.js"
+import Login from "../Pages/login.js";
+import ForgotPassword from "../Pages/resetPass.js";
+import SignUp from "../Pages/signup.js";
+import SignUpSupporter from "../Pages/signUpSupporter.js";
+import tos from "../Pages/TOS.js";
+import Home from "../Pages/home.js";
+import appts from "../Pages/appts.js";
+import Account from "../Pages/account.js";
+import Matching from "../Pages/search-appointment.js"
 import { Route, Switch, Redirect } from "react-router-dom";
 import UnauthenticatedRoute from "./UnauthenticatedRoute";
 import AuthenticatedRoute from "./AuthenticatedRoute";
